Lobby and handshaking code finished.
Lots of TODOs to fill in... haven't decided how to handle most of the
IOExceptions yet...

Where to from here? Probably a Message class to wrap either a game
move or chat message. OR more work on the GUI...

This commit could make a good place to fork from to make your own
game: implement LobbyListener and register your object, then call the
RemoteConnection's closeStatus().